Update entries for OpenSMC
The brigade member who set up our github account and web domain had a sudden emergency and was unable to pass authority to any other brigade members.  In the meantime, we've set up alternative resources to display on our section of the brigade site as we prepare for our first official hack night Jan 5, 2016!