Replace BMG Hack with Code for Bloomington
BMG Hack is in the midst of a naming transition toward Code for
Bloomington. Although their web properties all still call the
organization "BMG Hack" we can help them along by calling it Code for
Bloomington (BMG Hack) on our site -- the purpose of the parenthetical
being to not surprise the user if they click through to the BMG Hack web
page.