Add Open NC Collaborative
The data modeling here is a bit sketchy, because Open NC Collaborative
is not a brigade in the same way that the others are. But for now, let's
err on the side of treating it more like a brigade than not. (i.e.
applying the "Official" and "Brigade" tags to it)

I created a new tag to capture the two known instances of this -- the
other being CA Civic Lab (soon to be Code for California?).