Add `previous_names` to JSON and two brigades
There are a couple Brigades that have had inconsistent names across
different data sources. In order to support a data standardization
use-case and also to support Brigades changing name over time, let's add
an array of `previous_names` that can be used to get a bit of history
for Brigades. We will be using this on the CfA brigade website to
generate redirects for old Brigade pages.