Update brigade attributes to match CfA salesforce
This is mostly minor changes, a slash here, an "https" there... but
there are a couple actual changes:

For the following Brigades I've updated the `projects_list_url` to match
the Github account CfA has on file for the brigade. I did this only for
Brigades where the project list CSV did not load or it was out-of-date
in comparison to the Brigade's website. This affects the following
brigades:

* Code for ABQ
* Code for Boston
* Code for Durham
* Code for Hampton Roads
* Code for Sacramento
* Code for San Francisco

For those Brigades, the Projects list on the Brigade website will now be
updated from Github, rather than from the CSV. Please let me know if
this is in error.