Update `hack-for-la` projects_tag and documentation
The `projects_tag` feature was added so that Brigade leaders could
add a tag to projects in different organizations and have them all
appear on their Brigade's entry in the project index.

There was a discrepancy between this intended behavior and the actual
behavior:
* If `project_list_url` was specified, it would take precedence.
* The topic tag needed specification via full URL in order to match

This commit updates the README and Hack for LA's page to match this
behavior.